Move Chain and ChainState to own module
